chore: fix compatible issue with oz

+++ contracts/base/interface/uniswap/IUniswapV2Migrator.sol
@@ -1,14 +1,7 @@
 // SPDX-License-Identifier: MIT
 pragma solidity >=0.5.0;
 
-
 interface IUniswapV2Migrator {
-
-    function migrate(
-        address token,
-        uint amountTokenMin,
-        uint amountETHMin,
-        address to,
-        uint deadline
-    ) external;
+    function migrate(address token, uint256 amountTokenMin, uint256 amountETHMin, address to, uint256 deadline)
+        external;
 }
